Modify actor identification (TODO: Complete)

+++ mrpython/tincan/tracing_io.py
@@ -1,77 +1,21 @@
 """
-Modify the file "tracing_data.json" and keep a Json object with two keys:
-"student_hash": the hash used to identify a student
+Modify the file "backup_file" and keep a Json object with two keys:
 "statements": a stack of statements that couldn't be sent (no internet, LRS malfunction...)
 """
+from tincan import (tracing_config as config)
 import json
 import os
 
-filename = os.path.join(os.path.dirname(__file__), 'tracing_data.json')
-
-
-def reset_file():
-    with open(filename, "w") as f:
-        data = {"student_hash": "default", "student_context": "default", "list_statements" : []}
+backup_filepath = config.backup_filepath
+if not os.path.isfile(backup_filepath):
+    with open(backup_filepath, "w") as f:
+        data = {"statements": []}
         json.dump(data, f, indent=2)
-
-
-def modify_student_hash(student_hash, student_context):
-    with open(filename, "r+") as f:
-        data = json.load(f)
-        data["student_hash"] = student_hash
-        data["student_context"] = student_context
-        f.truncate(0)
-        f.seek(0)
-        json.dump(data,f, indent=2)
-
-
-def get_student_hash():
-    with open(filename, "r") as f:
-        data = json.load(f)
-        return(data["student_hash"])
-
-
-def modify_partner_hash(partner_hash):
-    with open(filename, "r+") as f:
-        data = json.load(f)
-        data["partner_hash"] = partner_hash
-        f.truncate(0)
-        f.seek(0)
-        json.dump(data,f, indent=2)
-
-
-def get_partner_hash():
-    with open(filename, "r") as f:
-        data = json.load(f)
-        return(data["partner_hash"])
-
-
-def modify_machine_id(partner_hash):
-    with open(filename, "r+") as f:
-        data = json.load(f)
-        data["machine_id"] = partner_hash
-        f.truncate(0)
-        f.seek(0)
-        json.dump(data,f, indent=2)
-
-
-def get_machine_id():
-    with open(filename, "r") as f:
-        data = json.load(f)
-        return(data["machine_id"])
-
-
-
-def get_student_context():
-    with open(filename, "r") as f:
-        data = json.load(f)
-        return(data["student_context"])
-
 
 def add_statement(statement):
     """Add a statement to the stack"""
     statement = statement.to_json()
-    with open(filename, "r+") as f:
+    with open(backup_filepath, "r+") as f:
         data = json.load(f)
         list_statements = data["statements"]
         list_statements.append(statement)
@@ -82,7 +26,7 @@ def add_statement(statement):
 
 def get_statement():
     """Get the first statement of the stack"""
-    with open(filename, "r") as f:
+    with open(backup_filepath, "r") as f:
         data = json.load(f)
         list_statements = data["statements"]
         if list_statements:
@@ -92,7 +36,7 @@ def get_statement():
 
 def remove_statement():
     """Remove the first statement of the stack"""
-    with open(filename, "r+") as f:
+    with open(backup_filepath, "r+") as f:
         data = json.load(f)
         if data["statements"]:
             data["statements"].pop(0)
